+++ packages/atlas/src/views/studio/VideoWorkspace/VideoForm/VideoForm.tsx
@@ -60,8 +60,8 @@ import {
 } from './VideoForm.styles'
 
 const CUSTOM_LICENSE_CODE = 1000
-const SCROLL_TIMEOUT = 700
-const MINT_NFT_TIMEOUT = 1200
+const SCROLL_TIMEOUT = 500
+const MINT_NFT_TIMEOUT = 800
 const MIN_TITLE_LENGTH = 3
 const MAX_TITLE_LENGTH = 60
 const knownLicensesOptions: SelectItem<License['code']>[] = knownLicenses.map((license) => ({
@@ -307,7 +307,7 @@ export const VideoForm: FC<VideoFormProps> = memo(({ onSubmit, setFormStatus }) 
       SCROLL_TIMEOUT
     )
     const setMintNftTimeout = setTimeout(
-      () => setValue('mintNft', tabData.mintNft || mintNft, { shouldTouch: true }),
+      () => setValue('mintNft', tabData.mintNft || mintNft, { shouldDirty: true }),
       MINT_NFT_TIMEOUT
     )